+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Sort grazing events chronologically.
 - Do not allow the same movement log to be used in multiple grazing plans.
+- Do not allow movement logs with multiple assets/locations.
 
 ### Fixed